create and move subcomponents
